requirements.txt
langchain
langchain-openai
langchain-google-genai
python-dotenv
langchain-community
tiktoken
lark
redis


